# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Darbby Platform API — Backend for customer and merchant apps
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
  /auth/user/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/merchant/register:
    post:
      operationId: registerMerchant
      tags: [auth]
      summary: Register a new merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantRegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/merchant/login:
    post:
      operationId: loginMerchant
      tags: [auth]
      summary: Merchant login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ─── VEHICLES ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UserRegisterRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
      required: [fullName, email, password]
    MerchantRegisterRequest:
      type: object
      properties:
        businessName:
          type: string
        ownerName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
        commercialRegNo:
          type: string
        commercialRegDocUrl:
          type: string
        commercialRegExpiry:
          type: string
          format: date
      required: [businessName, ownerName, email, phone, password, commercialRegNo, commercialRegDocUrl, commercialRegExpiry]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        actor:
          type: string
          enum: [USER, MERCHANT]
        id:
          type: string
        email:
          type: string
        name:
          type: string
      required: [token, actor, id, email, name]

    # ─── Vehicle ─────────────────────────────────────────────────────────────
...